+++ backend/utilities.js
@@ -1,16 +1,19 @@
 const jwt = require('jsonwebtoken');
+const dotenv = require('dotenv')
+dotenv.config()
+const { ACCESS_TOKEN_SECRET } = process.env;
+
 
 const authenticationToken = (req, res, next) => {
-    const authHeader = req.headers['authorization'];
-    const token = authHeader && authHeader.split(' ')[1];
+    const token = req.headers["authorization"];
+    if (!token) return res.sendStatus(403);
 
-    if (!token) return res.sendStatus(401);
-
-    jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, (err, user) => {
+    jwt.verify(token, ACCESS_TOKEN_SECRET, (err, user) => {
         if (err) return res.sendStatus(403);
         req.user = user;
         next();
     });
 };
 
-module.exports = { authenticationToken };
+
+module.exports = authenticationToken ;
